+++ qt/scientific_interfaces/Inelastic/Common/CMakeLists.txt
@@ -12,6 +12,7 @@ set(SRC_FILES
     SettingsPresenter.cpp
     SettingsView.cpp
     IndirectTab.cpp
+    InterfaceUtils.cpp
     ValidationUtils.cpp
 )
 
@@ -22,6 +23,7 @@ set(INC_FILES
     OutputPlotOptionsPresenter.h
     SettingsHelper.h
     SettingsPresenter.h
+    InterfaceUtils.h
     ValidationUtils.h
 )
